+++ packages/patternfly-4/react-core/src/components/ClipboardCopy/ClipboardCopyToggle.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import ToggleButton from './ToggleButton';
+import { ClipboardCopyToggle } from './ClipboardCopyToggle';
 
 const props = {
   id: 'my-id',
@@ -13,20 +13,20 @@ const props = {
 
 test('toggle button render', () => {
   const desc = 'toggle content';
-  const view = shallow(<ToggleButton {...props} aria-label={desc} />);
+  const view = shallow(<ClipboardCopyToggle {...props} aria-label={desc} />);
   expect(view).toMatchSnapshot();
 });
 
 test('toggle button onClick', () => {
   const onclick = jest.fn();
-  const view = shallow(<ToggleButton {...props} onClick={onclick} />);
+  const view = shallow(<ClipboardCopyToggle {...props} onClick={onclick} />);
   view.find('button').simulate('click');
   expect(onclick).toBeCalled();
 });
 
 test('toggle button is on expanded mode', () => {
-  let view = shallow(<ToggleButton {...props} isExpanded />);
+  let view = shallow(<ClipboardCopyToggle {...props} isExpanded />);
   expect(view.props()['aria-expanded']).toBe(true);
-  view = shallow(<ToggleButton {...props} isExpanded={false} />);
+  view = shallow(<ClipboardCopyToggle {...props} isExpanded={false} />);
   expect(view.props()['aria-expanded']).toBe(false);
 });