+++ packages/patternfly-4/react-topology/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.8...@patternfly/react-topology@2.4.9) (2019-06-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.7...@patternfly/react-topology@2.4.8) (2019-06-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.6...@patternfly/react-topology@2.4.7) (2019-06-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.5...@patternfly/react-topology@2.4.6) (2019-06-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
 ## [2.4.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.4...@patternfly/react-topology@2.4.5) (2019-06-25)
 
 **Note:** Version bump only for package @patternfly/react-topology